refactor: StoreSearchCondition을 application 계층으로 이동
