+++ src/main/java/com/sparta/delivery/store/domain/repository/StoreRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.sparta.delivery.store.domain.repository;
 
+import com.sparta.delivery.store.application.dto.StoreSearchCondition;
 import com.sparta.delivery.store.domain.entity.Store;
-import com.sparta.delivery.store.presentation.dto.StoreSearchCondition;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
