+++ src/main/java/com/sparta/delivery/store/infrastructure/persistence/repository/StoreRepositoryImpl.java
@@ -7,9 +7,9 @@ import static com.sparta.delivery.store.domain.entity.QStore.store;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.sparta.delivery.store.application.dto.StoreSearchCondition;
 import com.sparta.delivery.store.domain.entity.Store;
 import com.sparta.delivery.store.domain.repository.StoreRepositoryCustom;
-import com.sparta.delivery.store.presentation.dto.StoreSearchCondition;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
