+++ src/test/java/com/sparta/delivery/store/presentation/StoreControllerTest.java
@@ -22,9 +22,9 @@ import com.sparta.delivery.common.config.security.SecurityConfig;
 import com.sparta.delivery.common.response.PageResponse;
 import com.sparta.delivery.common.config.security.UserPrincipal;
 import com.sparta.delivery.store.application.StoreService;
+import com.sparta.delivery.store.application.dto.StoreSearchCondition;
 import com.sparta.delivery.store.presentation.dto.StoreCreateRequest;
 import com.sparta.delivery.store.presentation.dto.StoreResponse;
-import com.sparta.delivery.store.presentation.dto.StoreSearchCondition;
 import com.sparta.delivery.store.presentation.dto.StoreUpdateRequest;
 import com.sparta.delivery.user.application.UserService;
 import com.sparta.delivery.user.domain.entity.UserRole;
